requirements.txt
Flask==2.2.3
gunicorn==20.1.0
supabase
pyyaml
flask_cors